requirements.txt
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
curl_cffi==0.11.4
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
frozendict==2.4.6
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multitasking==0.0.11
numpy==2.3.1
pandas==2.3.0
peewee==3.18.1
platformdirs==4.3.8
protobuf==6.31.1
pycparser==2.22
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
six==1.17.0
soupsieve==2.7
SQLAlchemy==2.0.41
tabulate==0.9.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
waitress==3.0.2
websockets==15.0.1
Werkzeug==3.1.3
yfinance==0.2.64
